Fixes @yaatt/utils tests and adds build system
