+++ .gitignore
@@ -2,6 +2,5 @@
 .vscode
 
 node_modules
-dist
 *.map
 *.log
