+++ packages/yaatt-core/src/Request.js
@@ -4,7 +4,7 @@ import { pick, compose, prop, evolve, mergeDeepRight, map, chain } from 'ramda';
 import Future from 'fluture';
 
 import { toParams, mapFutureAsync, request, tryF, mapToList, listToMap, constant, createClass, validateRequest } from '@yaatt/utils';
-import Response from './Response';
+import { Response } from './Response';
 
 const callDependency = ({ key, value }: Object): Future =>
 	Request(value)
@@ -80,4 +80,4 @@ const Request = createClass({
 	),
 });
 
-export default Request;
+export { Request };
