+++ packages/yaatt-core/tests/Request.test.js
@@ -1,6 +1,6 @@
 import {identity} from 'ramda';
 
-import Request from '../Request';
+import Request from '../src/Request';
 import { request } from '@yaatt/utils';
 
 describe('Request', () => {
