+++ packages/yaatt-core/tests/test-runner.test.js
@@ -1,7 +1,7 @@
 
 // import Future from 'fluture';
 
-import { runTestCase, runTestSuite } from '../test-runner';
+import { runTestCase, runTestSuite } from '../src/test-runner';
 import { konsole } from '@yaatt/utils/logger';
 import { request } from '@yaatt/utils';
